# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Aegis Rise member, social feed, and chapter administration API.
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/social/connect/{platform}:
    post:
      operationId: createSocialConnection
      tags: [social]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SocialPlatform"
      responses:
        "200":
          description: A short-lived browser redirect URL for provider authorization
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialConnectResult"
    get:
      operationId: redirectToSocialProvider
      tags: [social]
      parameters:
        - $ref: "#/components/parameters/SocialPlatform"
        - name: state
          in: query
          required: true
          schema: { type: string }
      responses:
        "302":
          description: Redirect to the configured OAuth provider
  /auth/social/callback/{platform}:
    get:
      operationId: handleSocialCallback
      tags: [social]
      parameters:
        - $ref: "#/components/parameters/SocialPlatform"
        - name: code
          in: query
          schema: { type: string }
        - name: state
          in: query
          required: true
          schema: { type: string }
        - name: error
          in: query
          schema: { type: string }
      responses:
        "302":
          description: Redirect back to the member profile with the result
  /auth/social/accounts:
    get:
      operationId: listSocialAccounts
      tags: [social]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Connected social accounts without provider tokens
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SocialAccountList"
  /auth/social/disconnect/{platform}:
    delete:
      operationId: disconnectSocialAccount
      tags: [social]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SocialPlatform"
      responses:
        "200":
          description: Account disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
    get:
      operationId: disconnectSocialAccountGet
      tags: [social]
      security: [{ bearerAuth: [] }]
      parameters:
        - $ref: "#/components/parameters/SocialPlatform"
      responses:
        "200":
          description: Account disconnected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    SocialPlatform:
      name: platform
      in: path
      required: true
      schema: { $ref: "#/components/schemas/SocialPlatform" }
  schemas:
...
    Message:
      type: object
      required: [message]
      properties: { message: { type: string } }
...
    SocialPlatform:
      type: string
      enum: [facebook, linkedin, instagram]
    SocialConnectResult:
      type: object
      required: [authorizationUrl, expiresAt, provider]
      properties:
        authorizationUrl: { type: string, format: uri }
        expiresAt: { type: string, format: date-time }
        provider: { type: string }
    SocialAccount:
      type: object
      required: [platform, connectedAt, isActive, isPublishingEligible]
      properties:
        platform: { $ref: "#/components/schemas/SocialPlatform" }
        connectedAt: { type: string, format: date-time }
        expiresAt: { type: [string, "null"], format: date-time }
        isActive: { type: boolean }
        isPublishingEligible: { type: boolean }
        publishingError: { type: [string, "null"] }
    SocialAccountList:
      type: object
      required: [accounts]
      properties:
        accounts:
          type: array
          items: { $ref: "#/components/schemas/SocialAccount" }
...