# lib/api-zod/src/generated/types/socialAccount.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Aegis Rise member, social feed, and chapter administration API.
 * OpenAPI spec version: 0.1.0
 */
import type { SocialPlatform } from './socialPlatform';

export interface SocialAccount {
  platform: SocialPlatform;
  connectedAt: Date;
  /** @nullable */
  expiresAt?: Date | null;
  isActive: boolean;
  isPublishingEligible: boolean;
  /** @nullable */
  publishingError?: string | null;
}